Feat: Adds paths argument to lint_r
I had to create lint_path() that will switch either to
lintr::lint() for files and rhino:::lint_dir() for
directories. This is working in theory for now. My tests
wil files and directories looked ok. The problem with
testing it is the call to read_config() that requires
I run the package in a rhino directory. This makes me
wonder if I can test this function inside testthat. The
trick I used here requires purrr's map and reduce. It took
me a while to figure this out, but again, not sure if this
will work inside a rhino directory yet. I set the default
paths to NULL and assigned the app and tests/testthat
when the path is NULL so that lint_r() will still behave
the same after this update.