Feat: Maps lint_path with c
My initial process of map linting did not work as expected.
The lint errors were getting printed out after the map, and
no error messages were getting thrown. I realized that the
lint_dir call of the original lint_r wraps the lints in
c(). I included c to the lint_path map to prevent the
messages from appearing and to capture the lints. I still
had to use c with reduce because the output of map is
always a list. Fortunately, it worked.