Feat: Adds check if file or dir exists
I tried running lint_r() on a non-existent file/direcotry
I got this error:

Error in class(lints) <- "lints" : attempt to set an attribute on NULL

Not very informative to the user. Since the file/directory
does not exist, linting will return a NULL object. This
error happens when we try to assign full relative paths
to the lints object. I added a check to make sure that
the directory/file exists. Otherwise, the user will
receive a message that there is nothing to lint and
the specified path does not exist.