Feat: Adds check_paths function
My intial attempt at checking if the paths exist first
failed because of this lint_path().

When the path does not exist, it will do nothing and
return a NULL object. The error of assigning a lint
class to a NULL object then appears instead of my
abort message.

However, I did realize that I could not abort when a path
does not exist because that could complicate stuff when
a combination of existing and missing paths are provided
by the user.

I created the check_paths() function to return an error
when all the provided paths are missing, and remove
the missing paths and only keep the exisitng paths
before linting begins.

I did this because I did put a bad path by mistake, and the
error was confusing and not very informative to me.