Fix: Changes cypress paths in CI template
Changes working directory for cypress from .rhino/node to
.rhino, and project path from root/tests to ../tests.
All node stuff are now in .rhino and not in .rhino/node.
When cypress run tests, cypress looks for cypress.json
in .rhino/root/tests when it should look at the ./tests.
Cypress seems to be still stuck in .rhino, so it needs
to move a directory higher (..) to locates tests/ and
cypress.json.