fix: avoid double search box on Rhino home page
The `pkgdown/build.R` script used to replace the `navbar.html` template
in order to add a documentation version switcher to the navbar.
However, this solution broke down when the template was updated in {pkgdown}.

Fix: use `template.includes.before_navbar` to inject the version switcher
instead of overwriting the navbar template.
This solution has a chance of being more robust
against future updates to the template.